+++ api/src/main/java/com/jhsfully/api/model/permission/PermissionResponse.java
@@ -1,6 +1,8 @@
 package com.jhsfully.api.model.permission;
 
 import java.util.List;
+
+import com.jhsfully.api.model.dto.PermissionDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
